+++ fluss-client/src/main/java/org/apache/fluss/client/tiering/TableBucketWriteResult.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.fluss.flink.tiering.source;
+package org.apache.fluss.client.tiering;
 
 import org.apache.fluss.lake.writer.LakeWriter;
 import org.apache.fluss.metadata.TableBucket;
