+++ fluss-client/src/main/java/org/apache/fluss/client/tiering/TieringCommitResult.java
@@ -1,0 +1,52 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.fluss.client.tiering;
+
+import org.apache.fluss.lake.committer.TieringStats;
+
+import javax.annotation.Nullable;
+
+/**
+ * The result of one table's commit round, holding the lake committable (nullable for empty commits
+ * where no data was written) and the associated tiering statistics.
+ *
+ * @param <Committable> the type of the lake committable
+ */
+public class TieringCommitResult<Committable> {
+
+    /** The lake committable, or {@code null} if nothing was written in this round. */
+    @Nullable private final Committable committable;
+
+    /** Per-table tiering statistics collected during this round. */
+    @Nullable private final TieringStats stats;
+
+    public TieringCommitResult(@Nullable Committable committable, @Nullable TieringStats stats) {
+        this.committable = committable;
+        this.stats = stats;
+    }
+
+    @Nullable
+    public Committable getCommittable() {
+        return committable;
+    }
+
+    @Nullable
+    public TieringStats getStats() {
+        return stats;
+    }
+}
