+++ fluss-flink/fluss-flink-common/src/main/java/org/apache/fluss/flink/tiering/committer/CommittableMessageTypeInfo.java
@@ -17,8 +17,8 @@
 
 package org.apache.fluss.flink.tiering.committer;
 
+import org.apache.fluss.client.tiering.TableBucketWriteResult;
 import org.apache.fluss.flink.adapter.TypeInformationAdapter;
-import org.apache.fluss.flink.tiering.source.TableBucketWriteResult;
 import org.apache.fluss.lake.serializer.SimpleVersionedSerializer;
 
 import org.apache.flink.api.common.typeinfo.TypeInformation;
